requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.7
Django==3.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
